requirements.txt
alembic==1.11.1
anyio==3.6.2
arrow==1.2.3
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==7.1.2
cookiecutter==1.7.3
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.7.0
Mako==1.2.4
manage-fastapi==1.1.1
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
questionary==1.10.0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.14
starlette==0.27.0
text-unidecode==1.3
tomli==2.0.1
typer==0.4.2
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.12.2
wcwidth==0.2.6
